extend edn to include atom; tag 0.9.7
